requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
mysqlclient==2.2.8
pillow==12.1.1
PyJWT==2.11.0
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
